refactor: use a shared base class for duplicate init and from_csv
